unify status panels
