+++ app/views/jobs/_header.html.erb
@@ -5,4 +5,4 @@
   </div>
 </h1>
 
-<%= job_status_panel @job %>
+<%= status_panel @job %>
