+++ app/views/jobs/_job.html.erb
@@ -5,6 +5,6 @@
       <%= link_to "#{job.user.name} executed:", [@project, job] %>
       <pre class="pre-command"><%= job.command %></pre>
     </td>
-    <td><span class="label <%= deploy_status(job.status, 'label') %>"><%= job.status.titleize %></span></td>
+    <td><span class="label <%= status_label(job.status) %>"><%= job.status.titleize %></span></td>
   </tr>
 <% end %>
