+++ app/views/projects/_deploy.html.erb
@@ -8,6 +8,6 @@
       <%= content_tag :span, "Hotfix!", class: "label label-hotfix" if deploy.changeset.hotfix? %>
       <%= link_to "#{deploy.summary}", [@project || deploy.project, deploy] %>
     </td>
-    <td><span class="label <%= deploy_status(deploy.status, 'label') %>"><%= deploy.status.titleize %></span></td>
+    <td><span class="label <%= status_label(deploy.status) %>"><%= deploy.status.titleize %></span></td>
   </tr>
 <% end %>
